mkdir: revert default permissions from 0777 to 0755 (Resolves #15)
It is not desirable to create a mountpoint as world writable by default, as if
the filesystem is not mounted any user can write data that may be inadvertently
used by another service and security context.

Additionally, the directory assumes permissions from the NFS server once mounted
which will then be modified by this resource to make the entire NFS share world
writable on all servers.

This option is configurable and any user requiring this behavior can override
the default.

Documentation is also updated to reflect the new default, and clearly note that
owner, group and perm is propogated both to the mountpoint and remote server.

Partial revert of 968117be0926164c57eda5653b9bf2f7e1c8f5b3 (added gentoo compatibility)